Better split for missing-interpolator
