--- test/files/neg/t7848-interp-warn.check
@@ -1,33 +1,0 @@
-t7848-interp-warn.scala:20: warning: possible missing interpolator: detected interpolated identifier `$foo`
-    "An important $foo message!"                              // warn on ident in scope
-    ^
-t7848-interp-warn.scala:24: warning: possible missing interpolator: detected an interpolated expression
-    "A doubly important ${foo * 2} message!"                  // warn on some expr, see below
-    ^
-t7848-interp-warn.scala:27: warning: possible missing interpolator: detected interpolated identifier `$bar`
-  def i = s"Try using '${ "$bar" }' instead."                 // was: no warn on space test
-                          ^
-t7848-interp-warn.scala:28: warning: possible missing interpolator: detected interpolated identifier `$bar`
-  def j = s"Try using '${ "something like $bar" }' instead."  // warn
-                          ^
-t7848-interp-warn.scala:34: warning: possible missing interpolator: detected an interpolated expression
-  def v = "${baz}${bar}"                                      // warn on second expr
-          ^
-t7848-interp-warn.scala:35: warning: possible missing interpolator: detected an interpolated expression
-  def w = "${ op_* }"                                         // warn, only cheap ident parsing
-          ^
-t7848-interp-warn.scala:36: warning: possible missing interpolator: detected an interpolated expression
-  def x = "${ bar }"                                          // warn, a cheap ident in scope
-          ^
-t7848-interp-warn.scala:38: warning: possible missing interpolator: detected an interpolated expression
-  def z = "${ baz * 3}"                                       // warn, no expr parsing
-          ^
-t7848-interp-warn.scala:40: warning: possible missing interpolator: detected interpolated identifier `$this`
-  def thisly = "$this"
-               ^
-t7848-interp-warn.scala:41: warning: possible missing interpolator: detected an interpolated expression
-  def exprly = "${this}"
-               ^
-error: No warnings can be incurred under -Werror.
-10 warnings
-1 error
