--- test/files/neg/t9127.check
@@ -1,12 +1,0 @@
-t9127.scala:6: warning: possible missing interpolator: detected interpolated identifier `$s`
-  val t = "$s"
-          ^
-t9127.scala:7: warning: possible missing interpolator: detected an interpolated expression
-  val u = "a${s}b"
-          ^
-t9127.scala:8: warning: possible missing interpolator: detected interpolated identifier `$s`
-  val v = "a$s b"
-          ^
-error: No warnings can be incurred under -Werror.
-3 warnings
-1 error
