--- test/files/neg/t9127.scala
@@ -1,9 +1,0 @@
-// scalac: -Xlint:missing-interpolator -Xfatal-warnings
-//
-
-trait X {
-  val s = "hello"
-  val t = "$s"
-  val u = "a${s}b"
-  val v = "a$s b"
-}
